+++ src/collective/lazysizes/tests/test_setup.py
@@ -45,7 +45,7 @@ class InstallTestCase(unittest.TestCase):
         profile = 'collective.lazysizes:default'
         setup_tool = self.portal['portal_setup']
         self.assertEqual(
-            setup_tool.getLastVersionForProfile(profile), (u'1',))
+            setup_tool.getLastVersionForProfile(profile), (u'2',))
 
 
 class UninstallTestCase(unittest.TestCase):