+++ CHANGES.rst
@@ -4,6 +4,9 @@ Changelog
 1.0b1 (unreleased)
 ------------------
 
+- Use a blank image instead of a spinner as placeholder.
+  [hvelarde]
+
 - Logging now uses `debug` level instead of `info`.
   [hvelarde]